added catalog grid

+++ shop/models.py
@@ -5,7 +5,7 @@ class Product(DATABASE.Model):
     id = DATABASE.Column(DATABASE.Integer, primary_key = True)
     
     product_name = DATABASE.Column(DATABASE.String(50), nullable = False)
-    price = DATABASE.Column(DATABASE.Float, nullable = False)
+    price = DATABASE.Column(DATABASE.Integer, nullable = False)
     discount = DATABASE.Column(DATABASE.Float, nullable = False)
     count = DATABASE.Column(DATABASE.Float, nullable = False)
     description = DATABASE.Column(DATABASE.String(450), nullable = False)
